+++ src/main/java/kitchenpos/dao/MenuGroupDao.java
@@ -1,9 +1,8 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.MenuGroup;
-
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.MenuGroup;
 
 public interface MenuGroupDao {
     MenuGroup save(MenuGroup entity);
