+++ src/main/java/kitchenpos/dao/TableGroupDao.java
@@ -1,9 +1,8 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.TableGroup;
-
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.TableGroup;
 
 public interface TableGroupDao {
     TableGroup save(TableGroup entity);
