+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,6 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
 import java.math.BigDecimal;
+import kitchenpos.domain.Price;
 
 public class Product {
     private final Long id;
