+++ src/main/java/kitchenpos/product/ui/dto/ProductRequest.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui.dto;
+package kitchenpos.product.ui.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductRequest {
